skill/solana-security-audit-skill/pyproject.toml
[project]
name = "solana-security-audit-skill"
version = "0.1.0"
description = "OpenHarness skill that audits Solana/Anchor smart contracts."
readme = "SKILL.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
authors = [{ name = "SolGuard Contributors" }]

dependencies = [
    "anthropic>=0.34.0",
    "openai>=1.40.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "json-repair>=0.28.0",
]

[project.optional-dependencies]
# Installed via `uv sync --extra test` (kept as extra so end users who only
# want to run the skill don't need the pytest stack).
test = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
]

# `uv sync --extra parser` pulls the optional tree-sitter parser used in
# Phase 6 (§P2.2.2). Regex parser works without it.
parser = [
    "tree-sitter>=0.22.0",
    "tree-sitter-rust>=0.21.0",
]

[dependency-groups]
# Developer toolchain — activated via `uv sync` (default) or explicitly
# with `uv sync --group dev`. Not shipped as part of the runtime skill.
dev = [
    "ruff>=0.5.5",
    "black>=24.8.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "tools", "ai"]

# --- uv configuration ------------------------------------------------
# Pin Python via `.python-version` (3.11) and enable the default dev
# dependency group so `uv sync` yields a ready-to-hack environment.
[tool.uv]
default-groups = ["dev"]
managed = true

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["outputs", "work"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "S", "C90", "RUF"]
ignore = [
    "S101",
    "B008",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "N"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
disallow_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --strict-markers"
markers = [
    "live_llm: test requires a live LLM API key (skipped on CI by default)",
]
